fix(runtime): pass instanceRepo to NewRuntimeAgentHandler + round CPU cores to int
Companion to upstream #160 (cherry-pick 97bfbb5):

* runtime_agent_handler.NewRuntimeAgentHandler gained an instanceRepo
  parameter (used by syncInstanceRuntimeState for error binding cleanup).
  Wire the new argument at the call site in cmd/server/main.go.
* runtime_agent_request.CPUCores is int in the runtime agent's JSON
  schema; sending a float64 (e.g. 1.5 for 1500m cores) is rejected with
  "invalid json" and the gateway fails to start. Add cpuCoresToInt which
  rounds fractional cores up (0.5 → 1) so a partial-core instance still
  gets a non-zero cgroup quota.

+++ backend/internal/services/runtime_agent_client.go
@@ -32,7 +32,7 @@ type RuntimeAgentCreateGatewayRequest struct {
 	PortRange     RuntimeAgentPortRange `json:"port_range"`
 	UID           int                   `json:"uid"`
 	GID           int                   `json:"gid"`
-	CPUCores      float64               `json:"cpu_cores"`
+	CPUCores      int                   `json:"cpu_cores"`
 	MemoryMB      int                   `json:"memory_mb"`
 	DiskQuotaMB   int                   `json:"disk_quota_mb"`
 	Generation    int                   `json:"generation"`
